bytecode: changed method2WS syntax to method2(...).pop_unit to call methods which return nothing
This is a bit clumsy but it is the most generic and easy way it works
and will work for all types of method invocations. With method2WS and co. you
have to remember that you are calling a method with side-effects nonetheless,
so having the pop_unit syntax is more verbose but since it is generic I like
it more right now.